Property test for newState / events
They actually helped!
