+++ src/RenderingSpec.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE TemplateHaskell #-}
 module RenderingSpec where
 
@@ -8,6 +7,7 @@ import Test.Hspec
 import Test.Hspec.QuickCheck
 import Test.QuickCheck hiding (classes, once)
 
+import TreeGeneratorSpec
 import Component
 import Events
 import Rendering
@@ -16,29 +16,6 @@ data SingleConstructor = SingleConstructor
 
 $(deriveJSON (defaultOptions{tagSingleConstructors=True}) ''SingleConstructor)
 
-data TestPurviewT a m = Purview a m
-
-testHandler :: (String -> Purview String IO) -> Purview a IO
-testHandler = effectHandler ("" :: String) reducer
-  where
-    reducer :: String -> String -> IO (String, [DirectedEvent String String])
-    reducer action state = pure ("", [])
-
-testOnce = once (\send -> send "")
-
-sizedArbExpr :: Int -> Gen (Purview String IO)
-sizedArbExpr 0 = do pure $ text "always present"
-sizedArbExpr n = do
-  es <- vectorOf 2 (sizedArbExpr (n-1))
-  elements
-    [ div es
-    , style "" $ div es
-    , testHandler (const $ div es)
-    , testOnce (div es)
-    ]
-
-instance Arbitrary (Purview String IO) where
-  arbitrary = resize 3 $ sized sizedArbExpr
 
 spec :: SpecWith ()
 spec = parallel $ do
